refactor(api): separate app build and server start
