+++ test/routes/league.test.ts
@@ -1,5 +1,14 @@
 import test from 'ava';
 
-test('Should do a test on the league route', (t) => {
+test('should return a 400 when the league name is not provided', (t) => {
   t.is(4, 4);
 });
+
+test('should return a 400 when the league description is not provided', (t) => {
+  t.is(4, 4);
+});
+
+test('should create a new league', (t) => {
+  t.is(4, 4);
+});
+
